worker/requirements.txt
annotated-types==0.7.0
coverage==7.7.1
exceptiongroup==1.2.2
iniconfig==2.1.0
kafka-python==2.1.2
packaging==24.2
pip==23.0.1
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pydantic-settings==2.8.1
pytest==8.3.5
pytest-mock==3.14.0
python-dotenv==1.0.1
setuptools==65.5.1
tomli==2.2.1
typing_extensions==4.12.2
wheel==0.45.1
